+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.1.0
+- Added support for Main levels.
+- Added support for Editor levels.
+- Fixed negative normal attempts bug.
+- Fixed bug where best practice run was not saved at attempt 1.
+
 # 1.0.2
 - Geode 4.0.1 support.